refactor transfer action
